Add ArrayQueue and ArrayStack
I added both of these implementations.
The only other thing I change at any point is the test case. You can test on this commit for it before the test case has been changed.

That change was simply so I could test with an earlier Java Version. Please use this commit to test with the unedited test cases.